Remove license for UncheckyAds
The FadeMind lists the MIT license, however that project only hosts
the UncheckyAds host list and did not create it. Instead, that list
is based on content from unchecky.com.

I was unable to determine the license of that content, and am removing
the license to prevent someone from accidentally violating the
actual license by using the content.